+++ src/edit/qt/editor/subwindow/tag_editor_bitmap_subwindow.hpp
@@ -43,6 +43,8 @@ namespace Invader::EditQt {
 
         QComboBox *sequence;
         QComboBox *sprite;
+        
+        bool monochrome;
 
         enum Colors {
             COLOR_ARGB,
@@ -58,6 +60,8 @@ namespace Invader::EditQt {
         static void set_values(TagEditorBitmapSubwindow *what, QComboBox *bitmaps, QComboBox *mipmaps, QComboBox *colors, QComboBox *scale, QComboBox *sequence, QComboBox *sprite, QScrollArea *images, std::vector<Parser::BitmapGroupSequence> *all_sequences);
         void refresh_data();
         void reload_view();
+        
+        void generate_colors_array(bool monochrome);
 
         QGraphicsView *draw_bitmap_to_widget(Parser::BitmapData *bitmap_data, std::size_t mipmap, std::size_t index, Colors mode, int scale, const std::vector<std::byte> *pixel_data);
         void highlight_sprite(std::uint32_t *data, std::size_t real_width, std::size_t real_height);